# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AirGrid geospatial intelligence demo API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Pollutant = typeof Pollutant[keyof typeof Pollutant];


export const Pollutant = {
  PM25: 'PM25',
  PM10: 'PM10',
  NO2: 'NO2',
  O3: 'O3',
} as const;

export type StationStatus = typeof StationStatus[keyof typeof StationStatus];


export const StationStatus = {
  active: 'active',
  delayed: 'delayed',
} as const;

export interface Station {
  id: string;
  name: string;
  area: string;
  latitude: number;
  longitude: number;
  status: StationStatus;
  lastUpdated: string;
  pm25: number;
  pm10: number;
  no2: number;
  o3: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  dataQualityScore: number;
}

export type MeasurementSource = typeof MeasurementSource[keyof typeof MeasurementSource];


export const MeasurementSource = {
  measured: 'measured',
} as const;

export interface Measurement {
  stationId: string;
  timestamp: string;
  pollutant: Pollutant;
  value: number;
  unit: string;
  source: MeasurementSource;
}

export interface StationDetail {
  station: Station;
  recentMeasurements: Measurement[];
}

export interface GridCell {
  latitude: number;
  longitude: number;
  value: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  dataQualityScore: number;
}

export type PredictionGridMethod = typeof PredictionGridMethod[keyof typeof PredictionGridMethod];


export const PredictionGridMethod = {
  idw: 'idw',
} as const;

export interface PredictionGrid {
  pollutant: Pollutant;
  unit: string;
  method: PredictionGridMethod;
  cells: GridCell[];
  generatedAt: string;
}

export type HotspotKind = typeof HotspotKind[keyof typeof HotspotKind];


export const HotspotKind = {
  spike: 'spike',
  persistent: 'persistent',
} as const;

export type HotspotSeverity = typeof HotspotSeverity[keyof typeof HotspotSeverity];


export const HotspotSeverity = {
  elevated: 'elevated',
  high: 'high',
  critical: 'critical',
} as const;

export interface Hotspot {
  id: string;
  area: string;
  latitude: number;
  longitude: number;
  pollutant: Pollutant;
  value: number;
  baseline: number;
  kind: HotspotKind;
  severity: HotspotSeverity;
  detectedAt: string;
}

export type AirQualitySummaryProvider = typeof AirQualitySummaryProvider[keyof typeof AirQualitySummaryProvider];


export const AirQualitySummaryProvider = {
  demo: 'demo',
} as const;

export interface AirQualitySummary {
  city: string;
  provider: AirQualitySummaryProvider;
  providerLabel: string;
  stationCount: number;
  measuredCount: number;
  estimatedCellCount: number;
  hotspotCount: number;
  averagePm25: number;
  maxPm25: number;
  updatedAt: string;
  hotspots: Hotspot[];
  positioningStatement: string;
}

export type HistoricalMeasurementSource = typeof HistoricalMeasurementSource[keyof typeof HistoricalMeasurementSource];


export const HistoricalMeasurementSource = {
  demo: 'demo',
} as const;

export interface HistoricalMeasurement {
  stationId: string;
  timestamp: string;
  pm25: number;
  pm10: number;
  source: HistoricalMeasurementSource;
}

export type ValidationStatusStatus = typeof ValidationStatusStatus[keyof typeof ValidationStatusStatus];


export const ValidationStatusStatus = {
  not_loaded: 'not_loaded',
  loaded: 'loaded',
} as const;

export interface ValidationStatus {
  status: ValidationStatusStatus;
  message: string;
  datasetLabel: string;
  /** @nullable */
  filename?: string | null;
  /** @nullable */
  rowCount?: number | null;
  /** @nullable */
  trainingRows?: number | null;
  /** @nullable */
  validationRows?: number | null;
  /** @nullable */
  mae?: number | null;
  /** @nullable */
  rmse?: number | null;
  /** @nullable */
  r2?: number | null;
}

export type ListMeasurementsParams = {
pollutant?: ListMeasurementsPollutant;
};

export type ListMeasurementsPollutant = typeof ListMeasurementsPollutant[keyof typeof ListMeasurementsPollutant];


export const ListMeasurementsPollutant = {
  PM25: 'PM25',
  PM10: 'PM10',
  NO2: 'NO2',
  O3: 'O3',
} as const;

export type GetPredictionGridParams = {
pollutant?: GetPredictionGridPollutant;
};

export type GetPredictionGridPollutant = typeof GetPredictionGridPollutant[keyof typeof GetPredictionGridPollutant];


export const GetPredictionGridPollutant = {
  PM25: 'PM25',
  PM10: 'PM10',
  NO2: 'NO2',
  O3: 'O3',
} as const;

export type ListHistoricalMeasurementsParams = {
stationId?: string;
};