# lib/api-zod/src/generated/types/gridCell.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AirGrid geospatial intelligence demo API
 * OpenAPI spec version: 0.1.0
 */

export interface GridCell {
  latitude: number;
  longitude: number;
  value: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  dataQualityScore: number;
}