# lib/api-zod/src/generated/types/historicalMeasurement.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AirGrid geospatial intelligence demo API
 * OpenAPI spec version: 0.1.0
 */
import type { HistoricalMeasurementSource } from './historicalMeasurementSource';

export interface HistoricalMeasurement {
  stationId: string;
  timestamp: string;
  pm25: number;
  pm10: number;
  source: HistoricalMeasurementSource;
}