# lib/api-zod/src/generated/types/measurement.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AirGrid geospatial intelligence demo API
 * OpenAPI spec version: 0.1.0
 */
import type { MeasurementSource } from './measurementSource';
import type { Pollutant } from './pollutant';

export interface Measurement {
  stationId: string;
  timestamp: string;
  pollutant: Pollutant;
  value: number;
  unit: string;
  source: MeasurementSource;
}