# lib/api-zod/src/generated/types/pollutant.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AirGrid geospatial intelligence demo API
 * OpenAPI spec version: 0.1.0
 */

export type Pollutant = typeof Pollutant[keyof typeof Pollutant];


export const Pollutant = {
  PM25: 'PM25',
  PM10: 'PM10',
  NO2: 'NO2',
  O3: 'O3',
} as const;