# lib/api-zod/src/generated/types/station.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AirGrid geospatial intelligence demo API
 * OpenAPI spec version: 0.1.0
 */
import type { StationStatus } from './stationStatus';

export interface Station {
  id: string;
  name: string;
  area: string;
  latitude: number;
  longitude: number;
  status: StationStatus;
  lastUpdated: string;
  pm25: number;
  pm10: number;
  no2: number;
  o3: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  dataQualityScore: number;
}