# lib/api-zod/src/generated/types/stationDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AirGrid geospatial intelligence demo API
 * OpenAPI spec version: 0.1.0
 */
import type { Measurement } from './measurement';
import type { Station } from './station';

export interface StationDetail {
  station: Station;
  recentMeasurements: Measurement[];
}