# lib/api-zod/src/generated/types/validationStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AirGrid geospatial intelligence demo API
 * OpenAPI spec version: 0.1.0
 */
import type { ValidationStatusStatus } from './validationStatusStatus';

export interface ValidationStatus {
  status: ValidationStatusStatus;
  message: string;
  datasetLabel: string;
  /** @nullable */
  filename?: string | null;
  /** @nullable */
  rowCount?: number | null;
  /** @nullable */
  trainingRows?: number | null;
  /** @nullable */
  validationRows?: number | null;
  /** @nullable */
  mae?: number | null;
  /** @nullable */
  rmse?: number | null;
  /** @nullable */
  r2?: number | null;
}